Clarify "the image" in comment.
The code in question has two image variables: 'image' and
codec->internal->image (or gav1Image). So it is not clear which image
variable "the image" refers to. I believe it refers to
codec->internal->image (or gav1Image). So change "the image" to "the
decoder's image". This also matches the subsequent code that sets
image->decoderOwnsYUVPlanes to AVIF_TRUE.