decoder->image can't be NULL after decoder has run
decoder->image is set to NULL only in avifDecoderCleanup(), which is
only called in avifDecoderDestroy() and at the beginning of
avifDecoderParse(). Therefore, after decoder->image is allocated in
avifDecoderParse(), decoder->image cannot be NULL.

In avif.h, instead of saying when decoder->image "can be NULL", say when
decoder->image "is invalid".