Pass avifProperty to avifParse*Property functions
Pass avifProperty * instead of avifDecoderData * to the
avifParse*Property() functions. This gets rid of the repeated
appearances of the expression data->properties.prop[propertyIndex].

Fix a typo in the comment for the hasUnsupportedEssentialProperty member
in the avifDecoderItem struct.

Move the members of avifProperty into a union because only one of them
is used, depending on the 'type' member.