Use libaom cpu_used 6 in "good quality" usage mode
libaom 2.0.0 added cpu_used 6 in "good quality" usage mode. Although in
libaom 2.0.0 cpu_used 6 in "good quality" usage mode is just a place
holder and is identical to cpu_used 5, we can start using it now and it
will have an effect when a new version of libaom is released.