Remove checks for the "av01" brand
https://aomediacodec.github.io/av1-isobmff/ says a file conformant to
that spec shall have the 'av01' brand among the compatible brands array
of the FileTypeBox. The 'av01' brand (as opposed to the 'av01' image
item type) is not mentioned in the AVIF spec.

Therefore, 'av01' may be present in the compatible brands array of an
AVIF file, but it should not be the major brand of an AVIF file. In
addition, the presence of 'av01' in the compatible brands array does not
imply the file is AVIF.

Also use the || operator to connect the checks for the "avif" and "avis"
brands.