Remove MAX_COMPATIBLE_BRANDS
Change the compatibleBrands field of the avifFileType struct to a
pointer that points to compatible_brands in the read-only input. It is
not necessary to copy the compatible_brands to a buffer in the
avifFileType struct as long as we can keep the read-only input valid
for the lifetime of the avifFileType struct.