Miscellaneous cleanup of src/codec_rav1e.c
In rav1eSupports400(), return the boolean expression minorVersion >= 4
directly at the end of the function.

In rav1eCodecEncodeImage(), move the declaration and initialization of
supports400 into the block that uses the variable, and replace the
boolean expression
  (codec->internal->chromaSampling != RA_CHROMA_SAMPLING_CS400) &&
  (image->yuvFormat != AVIF_PIXEL_FORMAT_YUV400)
with the equivalent expression
  image->yuvFormat != AVIF_PIXEL_FORMAT_YUV400.

Note: The two boolean expressions are equivalent because
  image->yuvFormat != AVIF_PIXEL_FORMAT_YUV400
implies
  codec->internal->chromaSampling != RA_CHROMA_SAMPLING_CS400
if 'alpha' is false.