Clean up the avifImageYUVAnyToRGBAnySlow function
The main change is to move the declaration of ptrU8 and ptrV8 (and their
uint16_t * casts) to the outside of the for 'i' loop because they are
invariant to 'i'. Also replace uPlane[(uvJ * uRowBytes) + k] with
ptrU8[k] and replace vPlane[(uvJ * vRowBytes) + k] with ptrV8[k].

Correct some minor comment errors.