Move ptrU8,ptrV8,ptrU16,ptrV16 out of i for loop
Move the declaration and initialization of ptrU8, ptrV8, ptrU16, and
ptrV16 out of the 'i' for loop because these four variables are loop
invariant.

The downside is that these four variables are only used in the YUV
4:4:4 case. But I think the benefit is greater. Also, we could use
ptrU8 and ptrV8 in the non-4:4:4 cases if we change their offsets to be
of a signed type.