Optimise gbrp encoding
In the identity reformat mode (as opposed to YUV coefficients
reformatting), the only operation needed is to deinterleave components
into multiple planes.

This reduces CPU usage of this function from 1.84% + 0.43% of the total
profiling time of avifenc, to 0.13% on my Thinkpad x280, and from 0.90%
to 0.02% on my Nintendo Switch.

I also included a non-SIMD codepath, in case someone is running on a
different architecture.