avifImageIdentity8ToRGB8ColorFullRange needs 4:4:4
Check if image->yuvFormat is AVIF_PIXEL_FORMAT_YUV444 before calling
avifImageIdentity8ToRGB8ColorFullRange() because
avifImageIdentity8ToRGB8ColorFullRange() assumes the YUV 4:4:4 format.

Note that the AV1 spec, Section 6.4.2 (Color config semantics), page 119
actually has this requirement:

    If matrix_coefficients is equal to MC_IDENTITY, it is a requirement
    of bitstream conformance that subsampling_x is equal to 0 and
    subsampling_y is equal to 0.

But not all AV1 decoders enforce this requirement. For example, dav1d
does not. So libavif needs to validate its assumptions when converting
YUV to RGB with the identity matrix.

Note: I wrote a merge request for dav1d to enforce this requirement of
bitstream conformance, but I am not sure if it will be accepted:
https://code.videolan.org/videolan/dav1d/-/merge_requests/1063