Remove the colorRange field from avifCodecInternal
Remove the colorRange field from the avifCodecInternal struct. Now that
AVIF_RANGE_FULL is defined as 1 rather than 0x80, we can just cast the
color_range field in the codec's image struct to the avifRange enum type
directly. It is not necessary to convert with the ternary ? : operator
and cache the conversion result in the avifCodecInternal struct.