Fix GCC -Wclobbered warnings
GCC 9.3.0 warns that the variables 'f' and 'rgbDepth' in avifPNGWrite()
might be clobbered by 'longjmp' or 'vfork'. I am not familiar with
'vfork', but I believe this is a false positive for 'longjmp' because
those two variables are not modified between the 'setjmp' and 'longjmp'
calls.

Fix the warnings by moving the relevant code back to where it was before
commit 6ce2d81653878e4bef5304495a2876232b1c515c.

I tested with the command:

  avifdec Chimera_10bit_cropped_to_1920x1008_with_HDR_metadata.avif output.png

and got the expected error message:

  Conversion to RGB failed: output.png

Fix https://github.com/AOMediaCodec/libavif/issues/326.