Miscellaneous minor comment fixes
include/avif/internal.h: Change "full box" to "whole box" to avoid being
misinterpreted as the FullBox type in ISO BMFF.

src/read.c: Describe how the ftypData field of avifDecoderData is used.
Move the Note for avifParse() to the outside of the function to make it
more prominent. Remove an extra "This".