Read fails if offset is past end of buffer or file
Change avifIOMemoryReaderRead() and avifIOFileReaderRead() to return the
AVIF_RESULT_IO_ERROR error if offset is past the end of the buffer or
file.

In avifIOMemoryReaderRead(), if offset is exactly at the end of the
buffer, set out->data equal to the end of the buffer rather than the
beginning of the buffer. This is just for convenience and should not
matter because out->size is set equal to 0.

Also restructure avifIOMemoryReaderRead() and avifIOFileReaderRead() to
make the two functions as similar as possible. This minimizes the
behavior differences between the two functions.

In src/read.c, validate offset against decoder->io->sizeHint (if
available) before calling decoder->io->read(). If offset is greater than
decoder->io->sizeHint, return the AVIF_RESULT_BMFF_PARSE_FAILED error,
because this means the offset value that we read or calculate from some
ISO BMFF box is invalid.