Cleanup related to avifCodecGetNextImageFunc
Declare the 'sample' input parameter as const.

The avifCodecGetNextImageFunc implementations can assume the 'sample'
input parameter is not NULL.

The for (;;) loop in aomCodecGetNextImage() needs to set 'sample' to
NULL after 'sample' has been passed to a successful aom_codec_decode()
call.

Delete comments like "Check if there are more samples to feed" and
"Feed another sample" because they describe the following check in the
original code:
 codec->internal->inputSampleIndex < codec->decodeInput->samples.count
which has been deleted.