Fix two bugs in avifDecoderReadItem()
The two bug fixes are:

1. Change the test item->idatID == 1 to item->idatID != 0.

2. In the idatBuffer case, consider extent->offset when setting up
offsetBuffer.

Also make three cleanup changes:

1. In the !singlePersistentBuffer case, the size of item->mergedExtents
just needs to be totalBytesToRead, which may be less than item->size if
partialByteCount is specified. This requires moving forward the code
that sets totalBytesToRead.

2. In a memcpy() call, the second argument could be &item->mergedExtents
(the address of the struct) rather than &item->mergedExtents.data (the
address of the first member of the struct).

3. Remove a bytesToRead != offsetBuffer.size check, which is redundant
with the bytesToRead > offsetBuffer.size check later. At that point, we
know bytesToRead >= offsetBuffer.size, so checking
bytesToRead != offsetBuffer.size is equivalent to checking
bytesToRead > offsetBuffer.size.