Miscellaneous cleanup changes for src/io.c
In the debug printf statements, use the PRIu64 format to print the
uint64_t parameters 'offset' and 'size'.

In avifIOFileReaderRead() and avifIOFileReaderRead(), assume reader->f
is not NULL. This is guaranteed by avifIOCreateFileReader().

In avifIOFileReaderRead(), make sure 'offset' is in the range of 'long'
before casting it to 'long'. 'offset' is of the uint64_t type. On a
32-bit platform and also 64-bit Windows, 'long' is 32 bits. So casting
'offset' to 'long' could be problematic.