Change prototypes of functions for memory avifIO
Change the prototypes of avifDecoderReadMemory() and
avifDecoderSetIOMemory() to represent the input memory buffer by a
buffer data pointer and buffer size, instead of an avifROData pointer.
This makes their prototypes consistent with the prototype of
avifIOCreateMemoryReader(), fixing
https://github.com/AOMediaCodec/libavif/issues/344.

Also add error checking to avifDecoderSetIOMemory() and
avifDecoderSetIOFile(): check for the failure of
avifIOCreateMemoryReader() and avifIOCreateFileReader(), respectively.