Remove unneeded libs and includes for svt and aom
The additional libraries for svt were apparently copied from the rav1e
section. ws2_32.lib is the Winsock library, which is unlikely to be used
in a video codec. Similarly -ldl is for functions like dlopen() and
dlsym(), which are not found in the svt source tree.

The ext/aom/build.avif pathname is incorrect. Even the correct pathname
ext/aom/build.libavif is not needed, because headers in that directory
are not public headers.