Set g_lag_in_frames to 1 if encoding single image
The default value of g_lag_in_frames in libaom's "good quality" usage
mode is 19. In libaom's lookahead structure (struct lookahead_ctx) there
is an array of frame buffers. The number of frame buffers in that array
is max(1, g_lag_in_frames) + 1. Therefore this change reduces the number
of frame buffers from 20 to 2. This results in significant saving in
memory consumption for a large image, such as a photo taken on a
smartphone (typical resolutions are 3264x2448 and 4032x3024).

This change is suggested by Ranjit Kumar Tulabandu.