New CMake options: AVIF_CODEC_AOM_DECODE, AVIF_CODEC_AOM_ENCODE
These allow for someone to only use/link against libaom for its encode or decode capabilities,
instead of being forced to link against both.

Also, I removed the Open codec implementations for SVT and rav1e and documented in the internal
header that Open is just for the decode path. (That function could probably use a rename.)

avifCodecVersions() now also hints at what each compiled-in codec is actually offering (enc and/or dec).