Expose chroma sample position in decoded images
Have codecs expose the chroma sample position from the sequence header
OBU of the AV1 bitstream in the decoded images.

Right now avifDecoderParse() exposes the chroma sample position from the
av1C property and this value is left unchanged in the decoded images. It
would be useful (or at least more consistent with the other fields in
avifImage) to also expose the chroma sample position in the sequence
header OBU, in case it differs from the value in the av1C property.