Clean up comments related to idat fox for an item
Clean up the comments for the search for the idat box associated with an
item.

Note: I find "this meta box" a little confusing in those two functions
because the subject of the two functions is an item, not a meta box. So
I changed "this meta box" to "the meta box".