Add AVIF_MATRIX_COEFFICIENTS_YCGCO (YCgCo) support
Add AVIF_MATRIX_COEFFICIENTS_YCGCO=8 (YCgCo) support to src/reformat.c.
Only full range is supported. Limited range is not supported.

The YCgCo color space has the following two differences:

1. The limited-full range adjustment is applied to R,G,B as opposed to
Y,Cb,Cr. In this aspect, it is the same as Identity.

2. For the RGB-to-YUV conversion, it is similar to normal YUV
conversion but uses different equations (its equations do not use the Kb
and Kr coefficients).

To reuse the current implementation for normal YUV conversion, YCgCo is
only supported with full range so that we don't need to deal with the
different limited-full range adjustment.

Note: Although R,G,B limited-full range adjustment is also used with
Identity, it is easier to handle because Identity's matrix is a no-op
and therefore it does not matter whether range adjustment is performed
before or after a no-op.

Fix https://github.com/AOMediaCodec/libavif/issues/418