Add the avifContainerDump() function
Use the avifContainerDump() function in the info() function in avifdec.c
to dump the container info instead of the first frame's info.

I noticed this issue while trying to debug monochrome images encoded
with libaom before commit 2172ed05169f08b20245221a7e95619a65884fd7.
The container yuvFormat was YUV400 but the first frame's yuvFormat was
YUV420. avifdec -i printed YUV420, so I could not see the container
yuvFormat was wrong. Obviously it would be best to dump both the
container info and every frame's info, but that could be confusing to
average users, or at least too verbose.