Improve message for avifInputReadImage failure
Improve the error message for avifInputReadImage failure. The current
message is "Cannot determine input file format", but that is not the
only reason for avifInputReadImage failure. For example,
avifInputReadImage() fails if the CICP in 'image' is not supported.

Apparently we try to address this issue by having avifJPEGRead() print
an error message if its avifImageRGBToYUV() call fails. But I was still
confused by the "Cannot determine input file format" message when I
passed an unsupported --cicp option to avifenc.