Miscellaneous cleanup for src/write.c
Declare the cellIndex parameter of avifEncoderDataCreateItem() as
uint32_t to be consistent with the type of cellIndex elsewhere.

Add parentheses around the expressions to be cast to uint8_t.

Use a simpler equivalent expression to calculate mdatEndSearchOffset.

Declare imageWidth and imageHeight right before first use.

Remove an unneeded uint32_t cast when assigning the return value of
avifRWStreamOffset() to chunkOffset, which is no longer of the uint32_t
type.