Simplify avifEncoderAddImageInternal conditionals
Simplify some conditionals in avifEncoderAddImageInternal().

Since we require all cellImages have the same width, height, and depth,
move the checks for invalid values of width, height, and depth outside
the for loop and perform those checks on just firstCell.

Check gridCount > 1 instead of whether gridColorID or gridAlphaID is
nonzero. gridCount > 1 is an equivalent condition and is used elsewhere
in the function.

In the single-image case, we only need to check if the alpha planes of
cellImages have fully opaque values if firstCell has an alpha plane. (If
firstCell does not have an alpha plane, then no cellImages have an alpha
plane.)