Fix off-by-one errors in err msg/code in avifyuv.c (#506)
The condition in which it is unsafe to increment driftPixelCounts[drift]
should be drift >= MAX_DRIFT, not drift > MAX_DRIFT, and the error
message should say "drift greater than or equal to MAX_DRIFT".

Declare the maxDrift variable in the scope it is used in.

Zero-initialize the avifRGBImage rgb with memset() to ensure all struct
members are initialized.

Set rgb.width to the actual width (a + 1) before the
avifRGBImageUnpremultiplyAlpha() and avifRGBImagePremultiplyAlpha()
calls, otherwise the two functions will access uninitialized pixels.

Use "return 1" to exit the test on failure in test mode 3. This is
equivalent to exit(1) inside the main function, and is used for
consistency with test mode 1.