Set codec-specific options for color or alpha only
Add a new syntax for codec-specific options so that they can be applied
to the color or alpha sub-image only.

1. <key>=<value> applies to both color and alpha (if present).
2. color:<key>=<value> or c:<key>=<value> applies only to color.
3. alpha:<key>=<value> or a:<key>=<value> applies only to alpha (if
   present).

In the syntax message of avifenc, describe the libaom codec-specific
options that should not be used with the alpha-subimage.

Remove the special handling of the film-grain-test and film-grain-table
options for the alpha sub-image.

Fix https://github.com/AOMediaCodec/libavif/issues/533.