Clean up the code that checks imageCountLimit (#549)
Add the avifGetSampleCountOfChunk() function for the code that was
duplicated.

Declare the imageCountLimit parameter of
avifCodecDecodeInputGetSamples() to have the uint32_t type (because
decoder->imageCountLimit has the uint32_t type) and replace the uint64_t
variable imageCount with the uint32_t variable imageCountLeft. It is not
necessary to use the uint64_t type in this code.