Check aom_codec_control(AOME_SET_CPUUSED) failure (#570)
The aom_codec_control(encoder, AOME_SET_CPUUSED, aomCpuUsed) call fails
if aomCpuUsed is greater than the maximum cpu-used value supported by
libaom. This should not happen if we set aomCpuUsed correctly, so this
check is intended to prevent future bugs when we attempt to use libaom's
new speed 9, which is not supported by libaom 2.0.x or older.