Simplify the calculation of A (alpha)
Simplify the calculation of A (alpha) in avifImageYUVAnyToRGBAnySlow().
Make it look like the code that calculates Y, U, V in the same function.
For example, replace the state->yuvChannelBytes > 1 test with the
image->depth == 8 test (and swap the if and else branches), and omit the
clamping of ptrA8[i] because it cannot possibly be greater than
yuvMaxChannel.