Cast uint8_t variable to uint32_t before << 24
In avifParseItemPropertyAssociation(), cast the uint8_t variable
'version' to uint32_t before performing the << 24 operation. Without the
cast, 'version' is automatically converted to 'int' by the C language,
so if the most significant bit of 'version' is 1, it will be shifted
into the sign bit of the 'int' value.

Fix https://crbug.com/oss-fuzz/33032.