Remove const from non-pointer function parameters
Using 'const' on non-pointer function parameters means the function
cannot modify the parameters. This could be useful in function
implementations, but should be avoided in function declarations.

C++ allows the function declaration and definition to differ in the
constness of the parameters. Unfortunately C doesn't allow this, even
though gcc and clang allow this as a language extension.