Remove const from non-pointer function parameters
Using 'const' on non-pointer function parameters means the function
cannot modify the parameters. This could be useful in function
implementations, but should be avoided in function declarations.

C++ allows the function declaration and definition to differ in the
constness of the parameters. For example, in add.h we can declare

    int Add(int a, int b);

and in add.cc we can define the function with this prototpye:

    int Add(const int a, const int b)
    {
        return a + b;
    }

Unfortunately C doesn't allow this, even though gcc and clang allow this
as a language extension.