Require srcWidth,srcHeight <= 16384 for ScalePlane
In most cases a srcWidth or srcHeight of 32767 pixels or less should be
safe for libyuv's ScalePlane() and ScalePlane_12() functions. However,
in some cases a srcHeight less than 32767 still results in integer
overflows. It is not practical to emulate the complicated logic in
libyuv's scale functions and detect integer overflows. Require srcWidth
and srcHeight be less than or equal to 16384 as a simple conservative
check to avoid integer overflows in libyuv.