Fix compilation with 1755 ≤ LIBYUV_VERSION < 1774
ScalePlane_12 only provide optimization for 2x scale up using
linear/bilinear filter (when scaling up, box and bilinear are the same),
and for the rest cases, it simply calls ScalePlane_16, which handles 2x
scale up just fine (slower though).

This optimization is thus optional and compilation can be allowed for
libyuv versions not providing it.

Fixes GH-781 (the function was introduced in libavif
https://github.com/AOMediaCodec/libavif/commit/46104d6db42ec2a94a7da6fc0693f24fa90a1b98).