Support local android builds for libgav1
When building libavif for android using the NDK, the underlying
AV1 codec can be built for multiple architectures with specific
optimizations. It is useful to build them as it is sometimes
necessary to generate a fat APK (an APK that contains the library
for multiple architectures).

Upate the CMake file to include an architecture sub-directory while
finding libgav1. Also add a shell script that will clone and build
libgav1 using the andoroid NDK for the architectures supported on
Android by default.