Add Android JNI Bindings
Add JNI Bindings that can be used by android apps for AVIF decoding.
There are three functions that are exposed:
isAvifImage() and getInfo() and decode().

The README.md in the subdirectory contains instructions about how
to build the JNI wrapper and use it in an android app.

The Java class API can be found by going through the JavaDoc comments
in AvifDecoder.java

The following are the files of interest that need to be reviewed:
 * AvifDecoder.java - The Java interface for libavif.
 * build.gradle - Gradle build file for the java code.
 * libavif_jni.cc - The C++ file that interfaces between java and
                    libavif.
 * CMakeLists.txt - CMake file for building libavif_jni.cc and libavif.

All the other files in this commit are boilerplate files or scripts
generated by Android Studio.