reformat: Support F16 Half Float conversion in avifRGBImage
Some platforms (Android for example [1]) support only the F16 Half
Float format for images with bits-per-pixel == 16. This PR adds
support for converting the RGBA pixels into half float format.

It does so by adding a setting variable to avifRGBImage and then
doing the conversion at the end of avifImageYUVToRGB. It also
invokes libyuv if available to do the conversion with SIMD
optimizations.

While there is potential to make this even faster (instead of
doing this conversion in the end, we could potentially do it while
we do the YUV -> RGB conversion), that is a much larger change and
the performance gain from that may or may not be worth the trade
off in complicating the code paths.

[1] https://developer.android.com/reference/android/graphics/Bitmap.Config#RGBA_F16