avifArrayPop() should zero the popped element (#842)
avifArrayPop() should zero the memory of the popped element. If we ever
push an element to the array, the memory of the new element will be all
zeros.

Assert that arr->count > 0 in avifArrayPop().

Also make sure we cast arr->elementSize to size_t when we multiply it by
another uint32_t variable.