Write ccst box in Sample Entry for animated images
Section 7.2.3 of the HEIF spec:
The CodingConstraintsBox shall be present in the sample description
entry for tracks with handler_type equal to 'pict'.

The values of the fields are chosen to be permissive rather than
figuring out the exact values by inspecting the bitstream which can be
too cumbersome. The values are as follows:
all_ref_pics_intra - 0 - all samples can use any type of reference.
intra_pred_used - 1 - intra prediction may or may not be used.
max_ref_per_pic - 15 - reserved value to indicate that any number of
                       reference images can be used.

Fixes the following Compliance Warden error:
[heif][Rule #12] Error: CodingConstraintsBox ('ccst') shall be present
once