Fix FillImageGrid() yuvFormat assertion for alpha (#854)
Also fix AVIF_PIXEL_FORMAT_NONE comment.
Also adapt avifgridapitest to cover this use case and >8 bit depths.

An avifImage with NONE as yuvFormat means there is no YUV pixel, but
alpha samples can still be present, which is the way tiles are
decoded. See how yuvFormat is left to NONE for alpha images in
aomCodecGetNextImage() and other codec implementations.