Handle the new AOM_IMG_FMT_NV12 enum in switch
Handle the new AOM_IMG_FMT_NV12 enum conditionally in a switch statement
in aomCodecGetNextImage(). This fixes a Clang -Wswitch-enum warning,
which is enabled by -Weverything.

Although the latest code on the main branch of libaom supports the NV12
image format, it is only supported as an input to the encoder. The
aomCodecGetNextImage() function is a decoder function, so it should
treat the NV12 image format as unsupported.