Set libaom-specific option -a tune=ssim by default
Note that we have recommended setting -a tune=ssim only for high quality
levels. This pull request sets -a tune=ssim by default for all quality
levels, because we have not determined empirically the quality level
above which -a tune=ssim is better than -a tune=psnr. Also, it is easier
to understand the default value of an option if it is independent of
other options.

Part of https://github.com/AOMediaCodec/libavif/issues/955.