Skip YUV<->RGB conversion tests for large images
The built-in implementation of avifImageYUVToRGB() and
avifImageRGBToYUV() is slow for large images. For an image of
width=15107, height=3328, depth=10, the avifImageYUVToRGB() and
avifImageRGBToYUV() calls take about 5 seconds each on a Linux
workstation. Since avif_decode_fuzzer is run 100 times with a 90 seconds
timeout, we need to skip the YUV<->RGB conversion tests for large
images.

It is not clear what the image size threshold should be, so we begin
with the largest threshold (as an integer multiple of 1024 * 1024) that
excludes this particular test image.

Bug: b/241457214