googletest.cmd: Pass -Dgtest_force_shared_crt=ON
Pass the -Dgtest_force_shared_crt=ON option to cmake so that GoogleTest
is compiled with the DLL version of the MSVC runtime library (MSVCRT) on
Windows. By default GoogleTest uses the static library version of the
MSVCRT on Windows. All the other libraries (libavif, libaom, dav1d,
etc.) use the the MSVCRT DLL by default, so we will get errors about
using both the MSVCRT DLL and static library when linking
GoogleTest-based tests if GoogleTest is compiled with the MSVCRT static
library.