Fix some MSVC warnings
are_images_equal.cc: Include <string> because this file uses
std::stoi().

avifincrtest_helpers.cc: The '<' comparison in the following for loop
has a signed/unsigned mismatch:
  for (uint32_t plane = 0; plane < (info.monochrome ? 1 : AVIF_PLANE_COUNT_YUV);

Fix the warning by declaring 'plane' as 'int'.

avifrgbtoyuvtest.cc: The first two parameters of GetPsnr() are of the
'double' type but we are passing 'int64_t' arguments, and MSVC warns
about possible loss of data in the conversion from 'int64_t' to
'double'. Cast the 'int64_t' arguments to 'double'.