Add CI for Android JNI Bindings
This will build the JNI bindings with each PR/push to ensure that
we don't break the Android JNI build.

It is quite fast (< 3 minutes) when run locally.

We could potentially add some tests as well and run them in an
emulator on the CI. But that is work for the future.