avifEncoder: Add the autoTiling option
The autoTiling option, if set to AVIF_TRUE, causes avifEncoder to ignore
the tileRowsLog2 and tileColsLog2 options and automatically choose
suitable tiling values.

Also add the --autotiling option to avifenc, which sets --tilerowslog2
and --tilecolslog2 automatically.

The current implementation uses as many tiles as allowed by the minimum
tile area requirement and imposes a maximum of 8 tiles.