testutil::ReadImage()
Use AvifImagePtr in avifmetadatatest.cc to fix C API  memory leaks.

The aviftest_helpers target must be linked to avif_apps.
Use a STATIC avif_apps library instead of an OBJECT library to avoid
issues related to the lack of property transitivity in OBJECT libs.
Use object dependencies for libjpeg and libpng so that include
directories are propagated.