Refactor test_cmd.sh
Add a third parameter containing the path to a temporary directory for
build systems that cannot use /tmp.
pushd to that directory for test environments that do not support
creating files in the default run folder. This is necessary for
testing files beginning with a dash '-'.

Rename variables. Sort test categories from common use cases
(lossy/lossless) to specific checks (metadata).

Avoid exits so that there is a single cleanup line.