Write 'auxi' box for animated images with alpha channel
When alpha channel is present on animated images, the 'auxi' box is
mandatory as it tells the decoder what kind of auxiliary track it
is.

The current decoder gets the value from the still image item
properties. But if a decoder ignores the still image fields and
treats the animated AVIF as a video sequence, it has no way of
knowing what kind of information the auxiliary track contains.

This patch resolves the following compliance warden error:
[heif][Rule #1] Section 7.5.3.1
The nature of the auxiliary track is announced by the AuxiliaryTypeInfoBox that
shall be included in the sample entry of the auxiliary track.