Allow clamped cells in avifEncoderAddImageGrid()
The right-most column and bottom-most row of cells given to
avifEncoderAddImageGrid() can be smaller than the others.
This will allow for arbitrarily sized images to be encoded as grids.
avifEncoderAddImageInternal() will pad those truncated cells with
border samples values because all AVIF cells must have the same
dimensions according to HEIF.
(Not to be confused with AV1 tiles.)

Refactor avifgridapitest.cc.